feat: allow for index.html expansion
When providing links it is common to abbreviate `path/to/index.html` as
`path/to/`, i.e. to omit `index.html`. This change refines the check to
take that case into consideration: a path to be considered valid, needs
to point to a _file_ or to a _directory containing index.html_.

Existing check would check if a path exists, regardless of if that path
was pointing to a directory or to a file.