fix: correctly treat absolute paths against `dir`
Absolute paths should be rooted from the `--dir` path, not against the
base directory of the examined file. For example, given file in
`$DIR/sub/a.html` with a link to `/other/b.html` and deadlinks invoked
with `--dir $DIR`, that link should be evaluated as `$DIR/other/b.html`,
not as `$DIR/sub/other/b.html`.